+++ app/containers/System/index.js
@@ -1,82 +1,19 @@
-import React, { useState, useEffect } from 'react';
-import gql from 'graphql-tag';
-import { useMutation, useLazyQuery } from '@apollo/react-hooks';
-import { notification } from 'antd';
+import React from 'react';
+import { Switch, Route, useRouteMatch } from 'react-router-dom';
+
 import { System as SystemPage } from '@tip-wlan/wlan-cloud-ui-library';
 
-const GET_OUI = gql`
-  query GetOui($oui: String!) {
-    getOui(oui: $oui) {
-      manufacturerAlias
-      manufacturerName
-      oui
-    }
-  }
-`;
+import Manufacturer from 'containers/System/containers/Manufacturer';
 
-const UPDATE_OUI = gql`
-  mutation UpdateOui($oui: String, $manufacturerAlias: String, $manufacturerName: String) {
-    updateOui(
-      oui: $oui
-      manufacturerAlias: $manufacturerAlias
-      manufacturerName: $manufacturerName
-    ) {
-      manufacturerAlias
-      manufacturerName
-      oui
-    }
-  }
-`;
 const System = () => {
-  const [returnedOUI, setOUI] = useState(null);
-  const [updateOUI] = useMutation(UPDATE_OUI);
-  const [searchOUI, { data }] = useLazyQuery(GET_OUI, {
-    onError: () => {
-      notification.error({
-        message: 'Error',
-        description: 'No matching manufacturer found for OUI',
-      });
-    },
-  });
-
-  useEffect(() => {
-    if (data && data.getOui) {
-      setOUI(data.getOui);
-    }
-  }, [data]);
-
-  const handleUpdateOUI = (oui, manufacturerAlias, manufacturerName) => {
-    updateOUI({
-      variables: {
-        oui,
-        manufacturerAlias,
-        manufacturerName,
-      },
-    })
-      .then(() => {
-        notification.success({
-          message: 'Success',
-          description: 'Alias successfully updated.',
-        });
-      })
-      .catch(() =>
-        notification.error({
-          message: 'Error',
-          description: 'Alias could not be updated.',
-        })
-      );
-  };
-
-  const handleSearchOUI = oui => {
-    searchOUI({ variables: { oui } });
-  };
+  const { path } = useRouteMatch();
 
   return (
-    <SystemPage
-      onSearchOUI={handleSearchOUI}
-      onUpdateOUI={handleUpdateOUI}
-      returnedOUI={returnedOUI}
-    />
+    <SystemPage>
+      <Switch>
+        <Route exact path={`${path}/manufacturer`} component={Manufacturer} />
+      </Switch>
+    </SystemPage>
   );
 };
